fixed errors
added error message consistency,
added more error checking
readded csh

+++ rm.go
@@ -6,6 +6,13 @@ import (
 )
 
 func Remove() {
+
+	if len(os.Args) != 2 {
+		fmt.Println("Invalid Operation!")
+		fmt.Println("Syntax: rm [file]")
+		os.Exit(1)
+	}
+
 	file := os.Args[1]
 	err := os.Remove(file)
 	if err != nil {
